# nexus-hub-clone/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus OS API - Knowledge Management + AI Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages/{pageId}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a page
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update a page
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageBody"
      responses:
        "200":
          description: Updated page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a page
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /pages/{pageId}/backlinks:
    get:
      operationId: getPageBacklinks
      tags: [pages]
      summary: Get pages that link to this page
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backlinks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
  /pages/{pageId}/graph:
    get:
      operationId: getPageGraph
      tags: [pages]
      summary: Get graph data (nodes + edges) for a page's connections
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
        - name: depth
          in: query
          schema:
            type: integer
            default: 2
      responses:
        "200":
          description: Graph data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphData"
...
  /pages/{pageId}/content:
    get:
      operationId: getPageContent
      tags: [blocks]
      summary: Get page content as TipTap JSON
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Page content JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageContent"
    put:
      operationId: savePageContent
      tags: [blocks]
      summary: Save page content (TipTap JSON)
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePageContentBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageContent"

  # ── AI PROVIDERS ─────────────────────────────────────────────────────────────
...
  /databases/{databaseId}:
    get:
      operationId: getDatabase
      tags: [databases]
      summary: Get a database schema
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Database"
    patch:
      operationId: updateDatabase
      tags: [databases]
      summary: Update database schema or view
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDatabaseBody"
      responses:
        "200":
          description: Updated database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Database"
  /databases/{databaseId}/rows:
    get:
      operationId: getDatabaseRows
      tags: [db-rows]
      summary: Get rows from a database
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
        - name: view
          in: query
          schema:
            type: string
            enum: [table, kanban, calendar, gallery, list]
        - name: filter
          in: query
          schema:
            type: string
          description: JSON filter expression
        - name: sort
          in: query
          schema:
            type: string
          description: JSON sort expression
        - name: groupBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Database rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseRowsResponse"
    post:
      operationId: createDatabaseRow
      tags: [db-rows]
      summary: Create a row in a database
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatabaseRowBody"
      responses:
        "201":
          description: Created row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseRow"
  /databases/{databaseId}/rows/{rowId}:
    patch:
      operationId: updateDatabaseRow
      tags: [db-rows]
      summary: Update a row
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
        - name: rowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDatabaseRowBody"
      responses:
        "200":
          description: Updated row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseRow"
    delete:
      operationId: deleteDatabaseRow
      tags: [db-rows]
      summary: Delete a row
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
        - name: rowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── SEARCH ───────────────────────────────────────────────────────────────────
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List available templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"

  # ── ACTIVITY ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Page:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [page, database, wiki, project, daily, canvas]
        icon:
          type: string
        coverUrl:
          type: string
        parentId:
          type: string
        workspaceId:
          type: string
        authorId:
          type: string
        isPublished:
          type: boolean
        isFavorite:
          type: boolean
        isArchived:
          type: boolean
        tags:
          type: array
          items:
            type: string
        excerpt:
          type: string
        wordCount:
          type: integer
        linkCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, type, workspaceId, authorId]
...
    UpdatePageBody:
      type: object
      properties:
        title:
          type: string
        icon:
          type: string
        coverUrl:
          type: string
        parentId:
          type: string
        isPublished:
          type: boolean
        isFavorite:
          type: boolean
        isArchived:
          type: boolean
        tags:
          type: array
          items:
            type: string
...
    PageContent:
      type: object
      properties:
        pageId:
          type: string
        content:
          type: object
          description: TipTap JSON document
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
      required: [pageId, content]
    SavePageContentBody:
      type: object
      properties:
        content:
          type: object
          additionalProperties: true
      required: [content]

    # ── GRAPH ─────────────────────────────────────────────────────────────────
    GraphData:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
      required: [nodes, edges]
    GraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        icon:
          type: string
        linkCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required: [id, label, type]
    GraphEdge:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        label:
          type: string
      required: [id, source, target]

    # ── BLOCK ─────────────────────────────────────────────────────────────────
...
    Database:
      type: object
      properties:
        id:
          type: string
        pageId:
          type: string
        name:
          type: string
        defaultView:
          type: string
          enum: [table, kanban, calendar, gallery, list]
        properties:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseProperty"
        createdAt:
          type: string
          format: date-time
      required: [id, pageId, name, defaultView, properties]
    DatabaseProperty:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [text, number, select, multi_select, date, checkbox, url, email, phone, relation, formula, rollup, person, file, created_time, updated_time]
        options:
          type: object
          additionalProperties: true
        order:
          type: integer
      required: [id, name, type, order]
    UpdateDatabaseBody:
      type: object
      properties:
        name:
          type: string
        defaultView:
          type: string
          enum: [table, kanban, calendar, gallery, list]
        properties:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseProperty"
    DatabaseRow:
      type: object
      properties:
        id:
          type: string
        databaseId:
          type: string
        properties:
          type: object
          additionalProperties: true
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, databaseId, properties]
    DatabaseRowsResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseRow"
        total:
          type: integer
        grouped:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/DatabaseRow"
      required: [rows, total]
    CreateDatabaseRowBody:
      type: object
      properties:
        properties:
          type: object
          additionalProperties: true
      required: [properties]
    UpdateDatabaseRowBody:
      type: object
      properties:
        properties:
          type: object
          additionalProperties: true
      required: [properties]

    # ── SEARCH ────────────────────────────────────────────────────────────────
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        previewUrl:
          type: string
        isBuiltIn:
          type: boolean
        pageType:
          type: string
      required: [id, name, description, category, isBuiltIn]

    # ── ACTIVITY ──────────────────────────────────────────────────────────────
...