# nexus-hub-clone/lib/api-zod/src/generated/types/createDatabaseRowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexus OS API - Knowledge Management + AI Platform
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDatabaseRowBodyProperties } from "./createDatabaseRowBodyProperties";

export interface CreateDatabaseRowBody {
  properties: CreateDatabaseRowBodyProperties;
}