# nexus-hub-clone/lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexus OS API - Knowledge Management + AI Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Template {
  id: string;
  name: string;
  description: string;
  category: string;
  icon?: string;
  previewUrl?: string;
  isBuiltIn: boolean;
  pageType?: string;
}